backend/requirements.txt
fastapi
uvicorn
python-dotenv
python-multipart
torch
diffusers
transformers
accelerate
opencv-python
controlnet_aux
numpy
pillow
ultralytics
